Change `type_at_position` to `type_at_location`
Summary:
I revived the old `hover_info_for_position` logic, simplifying it to
`type_at_position`, to support a customer who is trying to work around
gaps in the existing `types(...)` query.

But the interface of using position is not a good one; there are many
expressions where the narrowest spanning symbol at the start and end
position are different, and in fact sometimes neither of these is
the symbol for the location!

For example `x if condition else y` is an expression whose type
is the join of those of `x` and `y`, but using either the start
or (stop - 1) positions would give the types of either `x` or `y`,
neither of which is correct.

So it's a *big* improvement for this use case if the actual end user's
use case is indexed by location rather than position; in this case
what we really want is a typed AST but that's a big project we can't
prioritize at the moment.

The implementation of this is a bit hacky - the logic is pretty
complex involving functors and some confusing visitors so I really
want to avoid diving into it and rewiring things (which could take days)
as opposed to accepting more of a high-level hack (which takes hours).

As a result, this implementation comes from noting that:
- the code already found all expressions matching a position, and
  then searched for the narrowest one
- any symbol matching a location must also match the start position
- so, all we actually need to do on top of existing code is replace
  the filter for a narrowest symbol with a filter for an exact
  location match.

*Duplicate matches?*

At the OCaml type level I can't rule out the possibility of multiple
matches. I *think* this can't happen (all the expressions I can
think of that nest require extra characters... if there is a counter
example it is by definition not really human-readable), but if it does then
the current implementation will arbitrarily pick the first match.

Reviewed By: grievejia

Differential Revision: D59892998

fbshipit-source-id: d676bfe6177ada480af7e7d992503bc7cc57daf1