Remove the implementation from simulated signature select
Summary:
There was a TODO here - from a type- and call-resolution standpoint,
an overloaded function's implementation signature should be *completely*
irrelevant according to PEP 484.

Our logic in simulated_signature_select (which is part of the type order,
although it is distinct from the "full" signature select used in actual
call resolution in AttributeResolution) was injecting the implementation.

This gets rid of that injection; I verified that the "full" signature
selection code is already ignoring the implementation so no cleanup
is needed there.

As a bonus this should be a (tiny) perf win, since we do fewer
comparisons.

Reviewed By: yangdanny97

Differential Revision: D59942972

fbshipit-source-id: ac6827a70f1168aa275b75d763a10a7e74baa44c