Add test documenting a `types(...)` bug
Summary:
We discovered last week that `types(...)` is incorrectly reporting
the type of the iterator in a for statement.

I suspect the problem is that even though the definition -> expression
mapping in actual code is (as far as I know) 1:1 in the sense that two
expressions cannot occupy the exact same range of positions, Pyre tends to
generate a *lot* of synthetic expressions, mainly in preprocessing and CFG
construction, and this means Pyre can very easily it cases where our
mangled AST does indeed fail to uniquely map location to expression (in
which case the `types(...)` output is arbitrarily picking one and not useful).

I believe this particular bug is caused by the CFG code; this commit adds
a test case demonstrating the problem and I'll try to stack a fix.

Reviewed By: vladima

Differential Revision: D60009040

fbshipit-source-id: 5235fafcd5d364d7d711a9fca1d1c33034d9b4a6