Fix for-loop type lookup by tweaking CFG logic
Summary:
This diff renames `For.preamble` to `For.synthetic_preamble` because
I want it to be more obvious from the name that this function is not
returning a part of an existing AST structure but rather synthesizing
something that does not already exist; that's important to be able to
see quickly when trying to understand how Pyre mangles complex statements.

In addition, this diff modifies the function so that all generated
expressions and statements used a made-up location that spans from the
start of `target` to the stop of `iterator`. This means that the location
won't overlap with any of the original expressions.

This fixes a bug where the `types(...)` query had the wrong type at the
location of the iterator because it was arbitrarily picking one of about
5 generated expressions that all had the same location when forming the
That is important becuse if `types(...)` relies on producing a
`Location.t -> Type.t` mapping.

Now, the mapping has an extra location due to the synthetic expressions, but
this is not a big problem for callers since they will probably never actually
look up that location.

Note that this bug could very easily pop up elsewhere in either preprocessing
or CFG - any time we reuse a location when making an expression that has a
different type it could lead to problems. I don't think we have time to do
we should keep this in mind so that we can debug quicker if more issues pop up.

Reviewed By: vladima

Differential Revision: D60009039

fbshipit-source-id: d8de608aeda8694239fcd1d302c01e7e82477aa6