Update triggered partial sinks' call info to be the actual call sites
Summary:
Currently, triggered partial sinks are under the `Origin` frames, which
prevents the UI from showing the sink traces of the partial sinks. This diff
lets triggered partial sinks to be under the `CallSite` frames, which can be
expanded by UI to show the sink traces.

Below are some note-worthy implementation details.

# Update the type of `TriggeredSinkForCall`
This diff updates the mechanism that passes the triggered sink taint, which is
generated from the forward analysis, to the backward analysis.

**Problem.** The current mechanism is described as follows (at each call site):
1. Forward analysis: Retrieve the backward model of the callee.
2. Forward analysis: For each actual argument, if it is triggered, modify the
partial sink in its taint tree (from the model) into triggered partial sinks,
and add extra traces. The result is stored as a map from expressions to taint
trees.
3. Backward analysis: Join the taint trees from Step 2 with the backward model,
and `apply_call`.

This is not ideal, because Step 3 `apply_call` to the taint that is generated
from Step 2. `apply_call` not only updates the call info in the taint (which is
intended and correct), but also updates the kinds and frames in the taint,
which is *not* intended:
1. For example, when the callinfo is `CallSite`, its associated extra traces
would be set to none during `apply_call`, but we want to keep the extra traces
so that they can be shown in the UI.
2. As another example, `apply_call` is also applied to the triggered sink kind,
which results in moving the local transforms into global transforms. This is
not intended because Step 2 already sets the correct taint kind — we don’t want
to update the taint kind again.

The root cause is the data structure that is used during the pass —
`TriggeredSinkForBackward`, which is a map from expressions to taint tree. The
taint trees are restrictive, because there are only a few operations that can
be applied to them. For example, in order to keep the extra traces during
`apply_call`, we would need to add an extra flag to special-case this.

**Solution.** This diff proposes to simply store `TriggeredSinkForCall` for
each call site in `TriggeredSinkForBackward`, as opposed to storing a taint
tree for each expression. The benefits are,
1. The above problems can now be solved.
2. Move the mutation of the sink taint into the backward analysis, to keep it
clean -- it is more natural for the forward analysis to work with source taint.

The new mechanism works as follows:
1. Forward analysis: Store `TriggeredSinkForCall` for each call site.
2. Backward analysis: `apply_call` to the backward model.
3. Backward analysis: Update the taint in the results of Step 2 using the
information from Step 1, such as modifying partial sinks into triggered partial
sinks and adding extra traces.

# Use existing callinfo and frames when updating partial sinks into triggered ones
When creating triggered partial sink taint during the forward analysis, we now
use the callinfo and the frame from the sink model (of the callee), instead of
`Declaration` callinfo and an empty frame. We also do this when creating an
issue that ends up with a triggered partial sink taint -- we use the callinfo
and the frame from the sink taint.

Reviewed By: arthaud

Differential Revision: D59686204

fbshipit-source-id: 5a9bac3bdc942eae257cb51996eec46935d1d990