Add unit test of FunctionDefinitionEnvironment update
Summary:
I discovered recently that Pyre isn't correctly triggereing type checks
in certain cases; the problem appears to be that `FunctionDefinitions _`
reverse dependencies aren't triggered.

This commit adds unit test cases demonstrating that the problem
which I can produce on a large real project does not easily reproduce in a
toy project - tests show that the existing code is correctly triggering
FunctionDefinitionEnvironment updates both when the raw source changes and
also when a wildcard import changes.

Reviewed By: grievejia

Differential Revision: D60052271

fbshipit-source-id: 107882e2dce574427b0b77a9aa5dcbe4c2c7902d