special-case classes w/ EnumMeta as enums
Summary:
We have special handling for classes that extend `enum.Enum`, but we don't do the same when the class's metaclass extends `enum.EnumMeta`.

This diff extends our special enum typechecking to cover those metaclasses, which comes up in a conformance test.

Note:

In 3.11 `EnumMeta` was renamed to `EnumType`, but they should be aliases for each other. I'm not sure if I need to explicitly check both, since IDK what happens if one of them gets removed in the future.

Reviewed By: stroxler

Differential Revision: D59965041

fbshipit-source-id: a22270a85b0cb9df8954ee9b3e933f433e09fe34