fix typed dictionary base class bug
Summary:
This is a fairly weird one - when we preprocess typed dict declarations, we transform them into entirely different class declarations.

This transformation had a few bugs:
1. it only works if TypedDict is the first argument in the base class list
2. any other base class is stripped from the transformed class

This means that some of the inheritance typechecking for typed dicts does not work for classes which directly extend `TypedDict`.

Our existing test cases only checked when a class extends another user-defined typed dict class (which behaves as intended), not `TypedDict` itself.

This diff slightly cleans up that logic to fix those two bugs, and allows TypedDict to extend Generic. We're now able to detect `TypedDict` in other positions in the base class list, and we now include the base class list (minus TypedDict and total=True/False) in the transformed class.

In the future we should really try to clean this up more, but that's out of scope for this diff.

There's a bunch of conformance issues and missing features with typed dicts, so whoever picks that up might also want to take a look at this.

Reviewed By: stroxler

Differential Revision: D59986384

fbshipit-source-id: 3cb6223b4ba78a619fbe5b77644103c4bb30023e