add non-passing test cases for TypedDict del
Summary:
Required keys in typed dicts may not be deleted/unset. I've added a couple of tests and marked it with TODOs.

We probably want to add a case for the subscript expression, checking that the left side is a typed dict & the right side is a string literal. Then we can look up the field in the typed dict and see if it's required.

Required/NotRequired aren't implemented yet, so the logic would need to eventually be updated for that.

Reviewed By: stroxler

Differential Revision: D59986898

fbshipit-source-id: 71883d1c91ad47d92c837b1c93c6fdb45e6d81f6