add strict overrides flag
Summary:
This diff adds a flag that will eventually be used to toggle the PEP698 strict override checks added in D59398364. Currently it is unused, I'll hook it up and add a test in the next diff.

This is a lot of files to modify to add a single flag, and we definitely need to change this if we want to eventually add a flag for tuning each type of error.

I guess there isn't a way to spread ocaml record types into each other, so maybe in the future these could be composed somehow w/ a shared nested record that contains a boolean flag for each error code?

Reviewed By: kinto0

Differential Revision: D60078685

fbshipit-source-id: ae3aeec797dc4b255cd4b5728a4d907cd5618dd4