Change the type of Type.create so that it doesn't deal with variable declarations.
Summary:
We continue to detangle type aliases and variable aliases for PEP695.

This diff does more refactoring for RawAlias.t, which is the TypeAlias and VariableAlias variant. Our  ultimate goal is parsing type variables as variables, not type aliases.

After this diff, we will stack another diff so that in attributeResolution, we are working with variables rather than variable declarations.

Reviewed By: stroxler

Differential Revision: D60075938

fbshipit-source-id: 16126d0c2291793b8f77226a099dc3154c14634b