Fix walrus operator handling in refinement
Summary:
A walrus operator expression ought to refine as if the name being
bound were the refinement target; this is actually a reasonable use
case for refinement since you can embed a type guard boolean check
inside of many comprehension forms and thereby avoid a lot of boilerplate
refinement statements.

This commit modifies the refinement logic in Pyre to avoid directly pattern
matching on `Name` expressions; instead we always match the underlying form
and then determine the name by calling a helper function that is capable of
understanding "expressions that map to a name" more generally; at the
moment only walrus operator and actual `Name` expressions fit the bill but
the plumbing is general.

Reviewed By: migeed-z

Differential Revision: D60151182

fbshipit-source-id: 8664094da2a7873a37a4642416f2927c03b294fd