Remove poorly-chosen names in moduleTrackerTest
Summary:
It is not a good idea to put spaces in an OUnit test name, especially
at the end of the name, because the space is almost unprintable which
means the user won't easily see the test name when trying to rerun just
one test.

This commit strips all trailing spaces from test names.

Reviewed By: yangdanny97

Differential Revision: D60208401

fbshipit-source-id: 946cc24d11d0a2ae0dc58c220da753767066a5a2