Handle EINTR while waiting for forked job handler
Summary:
Lwt installs a SIGCHLD handler, which it uses to reap child processes that it
creates. Because a signal handler exists, when the forked child exits, the
worker receives a SIGCHLD signal. This happens during the waitpid, so we end up
with the EINTR exception.

I'm not sure why this only affecs Mac and not Linux. My only guess is that Linux
and Darwin just have different behavior around whether SIGCHLD interrupts wait.
I found some discussion online that reflects this, but it's not very conclusive.

In any event, it's a good idea to wrap this waitpid in a EINTR handler
regardless, and it fixes the issue on Darwin without affecting the behavior on
Linux.

Fixes https://github.com/facebook/pyre-check/issues/868

Reviewed By: grievejia

Differential Revision: D60198675

fbshipit-source-id: 1fccad44aa85fb479e7571a393cf3b85780e8d05