Exit from workers on uncaught exception
Summary:
The worker process does not do much -- it listens for jobs on a pipe, forks, and
waits for the forked process to exit. If any of these steps fail and raise an
exception, we want the worker to exit.

When I changed workers from spawned processes to forks, the behavior around
exception handling changed unintentionally. Before, an uncaught exception would
unwind to the process entry point and hit the OCaml runtime's uncaught exception
handler, which causes the process to exit. Now, the exception unwinds through
the stack including the fork call, and is caught by a server exception handler
that just logs. When this happens, the worker is "stuck" -- it never returns to
listening to the pipe, but it also doesn't exit, causing Pyre to "hang."

This diff restores the exit-on-error behavior by installing an exception handler
at the logically entrypoint of the worker process, and exiting on uncaught
exceptions.

The next diff addresses the source of the exception. This is a defense-in-depth
change that fixes our behavior on exceptions in general.

Fixes https://github.com/facebook/pyre-check/issues/868

Reviewed By: yangdanny97

Differential Revision: D60198676

fbshipit-source-id: f63ea3d1f9ec44d97945fd6a4290d26421a0c0c9