Do not depend on the parsing logic located in create_logic when parsing type variables
Summary:
The goal is to remove the logic inside `create_logic` for parsing type variables. This requires us to  use the variable declaration parser instead. The diff after this will allow us to delete the type variable parsing logic in `create_logic`.

There's one unittest change, hence the RFC tag. In this unit test, we get one more type error, which is `Invalid type parameters [24]: Generic type `Container` expects 1 type parameter.`. I believe that this is coming from a dummy type variable declaration for the self type which looks like this `typing.TypeVar("_Self_test_Container__", $parameter$bound = "test.Container")` and we can see that the bounds aren't instantiated which is why we get this error. This is the correct behavior for legitatamate type variables, which is why we don't want to silence this error entirely. But by this point, AFAIK, we don't know if the type variable is from the original program or from preprocessing.

Ideally, we want to eliminate preprocessing the self type. We should add a new type representing it in our type system and avoid adding those dummy type vars.  Otherwise, we would have to keep track of which variables came from where. I'm submitting this for review to get opinions on the best way to move forward.

Reviewed By: stroxler

Differential Revision: D60255581

fbshipit-source-id: 82d56c5bdbf8214bf10f3ae3aa9926a456c6954b