Do not allow type variable definitions to include other type variables
Summary:
For now, we will not allow this. We will rethink this point when it's time to implement PEP695 support. But regardless, the funciton we modified won't be used because we cannot mix legacy syntax with PEP695 syntax.

We added a unittest that demonstrates the error. Before this change, Pyre would have crashed on the unittest.

Reviewed By: yangdanny97

Differential Revision: D60427423

fbshipit-source-id: 77a511fcafcfad2961082d038a045cbcd8d044fb