Fix saved-state loading bug
Summary:
IGwww has been complaining that they aren't getting incremental updates
in callgraph data even though Sam recently put the call graph data in the
TypeEnvironment which should have fixed incremental bugs.

Turns out the problem is actually not
in pyre core at all but startup: the server startup code doesn't thread the
call graph flag properly in the case of loading a saved state - we always
create an `EnvironmentControls.t` with the right flag at the top level, but
then `ServerState.load` was recreating one that lacked the normal flags
from scratch.

The fix is to pass the controls down - controls should *only* ever be created
at the top level in `serverCommand.ml`. I did an `rg EnvironmentControls.create`
from the `server/` directory to verify that we never create one (except in
tests) after this diff.

This is probably indicative of a related structural issue, which
is that too many flags have wound up in `EnvironmentControls.t`; the current
state with environment controls is much better than what we had before,
but several of these flags are actually global to an analysis, and really belong
in `Configuration.Analysis.t` which would help us detect mismatches in saved
state; in particular, you'll get undefined behavior if you try to use saved state
with lazy module tracking (to be fair that doesn't make sense, but we wouldn't
currently catch it and give a good error).

Reviewed By: grievejia

Differential Revision: D60457611

fbshipit-source-id: b9e3405683f69111847bc5c3cc242996a348b0fd