Do not crash on splatted RHS in unawaitedAwaitableCheck
Summary:
Last week we got reports of Pyre hard-crashing (the user-facing behavior
is pretty bad - because of multiprocessing, when a worker dies the main
process cleans up while other workers continue happily chugging away, producing
many thousands of parser errors when they encounter an already cleaned-up
project).

The root cause was `zip_exn` being used in a place where, when Starred expressions
are used on the right-hand-side of a multi-target assign, the lengths could be
mismatched.

Without putting a chunk of time in that isn't merited by current priorities I'm
not sure how to model this (multi-target assignment is one of the messier things
to model at the type level), but at the very least we should just keep going,
producing false negatives in some cases. This commit does that; I filed a task
and recorded in a test the false negative that results for future prioritization.

Reviewed By: grievejia

Differential Revision: D60394863

fbshipit-source-id: 26c1d851b8856632da2994f0a1ae5e7954e98c17