recognize _magic_enum_attr as a property decorator
Summary:
Refer to https://fb.workplace.com/groups/pyreqa/permalink/7769076553182109/ for context/motivations

TLDR is that the `.value` and `.name` properties of an enum are typed as `Any` due to the decorator being unrecognized.

Based on typeshed, the `@_magic_enum_attr` eventually ends up meaning `property` via some convoluted mechanism (pre 3.11).

https://github.com/python/typeshed/blob/fa70187245751c6fd4d057534af7f9ae3e74a25f/stdlib/enum.pyi#L187

This diff makes it so that `@_magic_enum_attr` gets recognized as a property decorator, which seems to fix the issue.

Reviewed By: stroxler

Differential Revision: D60469070

fbshipit-source-id: 97211ca1ba74fddff2728d55eb602b05e65efaf6