Add a doc comment to checkCommand.ml
Summary:
Still chipping away at T132410158. In this case I'm trying to
emphasize how the command is invoked and communicates to the
frontend, which is what a new dev working on this code probably
most needs to understand.

I also specifically call out that `pyre` doesn't hit this code
path, which is often a source of confusion for folks onboarding.

Reviewed By: grievejia

Differential Revision: D60477279

fbshipit-source-id: 144629a708e1cc13346d3988c76ee68e6c42accb