Remove the unmatched partial sinks at the end of each backward analysis
Summary:
This is an optimization to stop propagating early the partial sinks that cannot
result in creating an issue. This is possible because in order for a
multi-source issue to exist, there must exist two partial sinks from two
different parameters at some call site.

Implementation-wise, this diff modifies the backward model at the end of the
backward analysis, by checking that for each parameter in the model that has a
partial sink, if there exists another parameter that has a matching partial
sink (i.e., a partial sink that originates from the same call site).

Reviewed By: alexkassil

Differential Revision: D60292844

fbshipit-source-id: 60651a348751b84149871ca8c5e3b9b10bb09365