Remove legacy parsing logic for type variable declarations
Summary:
Finally, we remove the type aliasing (incorrect) parsing logic, as we now parse type variables as type var declarations.

This means that we cannot use the old parsing logic in our testsuite anymore, so we update it to use the declaration parser.

Furthermore, we do a location insensitive compare by comparing the strings. This is not an ideal situation as there could be a better way to handle this. So we can revisit this later on how to do better.

For globalResolutionTest, we need to fix a unittest which is maked by a todo. We most likely have to not use parse_annotation for assert_constructor for this.

Reviewed By: yangdanny97

Differential Revision: D60478679

fbshipit-source-id: 0f1af45aabb5ea8de9f21eb036ab543634531467