Describe `serverCommand.ml`
Summary:
Still chugging away at T132410158. I think it's useful to not only
describe what this is but also how it relates to various user-facing
commands, as well as explicitly emphasize that `pyre code-navigation`
is *not* using this entrypoint since that's a likely point of confusion
for anyone looking at the code for the first time.

Reviewed By: grievejia

Differential Revision: D60566002

fbshipit-source-id: fd879a9e58d056be371ce43b8051a0f916c9eba2