Add tests verifying overload resolution for ReadOnly
Summary:
Since ReadOnly stubs rely heavily on overloads that specify `self`,
it is very important to know we select overloads reliably and there
was actually no unit test.

This commit adds two tests, verifying Pyre behavior when the order is
sensible (read-write needs to come first, because a read-write reference
is a subtype of a read-only reference) and when it is not.

Reviewed By: grievejia

Differential Revision: D60613171

fbshipit-source-id: fb6d25a4f34c5cb4b9a7791604a950e28c947f60