calculate extends_enum in class metadata
Summary:
this diff exposes an extends_enum flag in class metadata which denotes whether or not a class inherits from Enum, and exposes it to GlobalResolution and AttributeResolution

this allows some code sharing between typeCheck and AttributeResolution, and eliminates the need to call GlobalResolution.less_or_equal in the former.

ideally, making a shared is_enum flag and sharing the metaclass check as well would be nicer, but I don't think classSuccessorMetadataEnvironment is able to compute this.

Reviewed By: stroxler

Differential Revision: D60535053

fbshipit-source-id: a575c7586ff11a0e208c7362731a1d88deba701d