Thread a variable argument in attributeResolution
Summary:
As per our plan https://docs.google.com/document/d/1d1oYn5T0rJzCBaBR_vHJUakqWSUyDegLcltP4jXmzLk/edit#heading=h.9o7xrqp1vjae

we want to allow attributeResolution to accept context. In this diff, we enable this for resolve_literal which is a method that accepts a literal (and now context), and an expression and returns a type.

This is a no-op but we are threading context through the entire attribute resolution.

As we can see in the existing unittest, the literals being tested require no context which is why we are passing an empty map.

We also enable this for create_attribute. We will gradually continue with this plan till most of attribute resolution accepts context.

Reviewed By: yangdanny97

Differential Revision: D60801908

fbshipit-source-id: f7e59bcf583673e48f9e7ff5ad3847a334c879e4