Fix bug in weakening mutable literals in list of lists
Summary:
Fixes a bug that causes a spurious "incompatible variable type" error when making an annotated assignment of a list of lists. The issue was that weaken_mutable_literals was trying to resolve a union type against an expected list type, which doesn't work. I changed the code to individually resolve the types in the union.

This change allowed me to get rid of the `resolve_items_individually` option, which also improved the output for one of the `TypedDict` tests in weakenMutableLiteralsTest.

Reviewed By: samwgoldman

Differential Revision: D60947805

fbshipit-source-id: 1aadef70ab651285dfdfaf339912763f154acc5d