annotate NamedTuple generated _field attr as ClassVar
Summary:
This whole preprocessing needs to be removed or cleaned up at some point, but for now I need to update the behavior of the _fields attr we generate. In the NamedTuple base class it's a ClassVar, and the class we generate annotates it as an instance var.

In the following diff this type of override will be banned.

Reviewed By: samwgoldman

Differential Revision: D61019442

fbshipit-source-id: b9612db5dd454b9c703effe50d5cf859776be3bb