allow Any as base class
Summary:
Classes w/ "Any" as the base class are supposed to have duck typing behavior.

We don't implement the duck typing behavior yet, but this diff changes our typechecking to not emit errors when we see Any in the base class list.

We may be able to partially support this behavior by treating classes extending Any as a mock class, which suppresses certain classes of errors.

Reviewed By: migeed-z

Differential Revision: D61022594

fbshipit-source-id: 2eef2b6c070b589df82a67910e1cfdf06078f6bf