Simplify and document generic alias handling
Summary:
I got pretty confused trying to skim this code to understand if
and how it interacts with `replace_unbound_parameters_with_any`
in `AttributeResolution`.

I decided to make the logic more straight-line top to bottom and
more importantly to put a comment in explaining how this works
(because generic alias handling actually isn't the most obvious
thing even in the typing spec, and this code is the source of
truth for how we do it).

Reviewed By: grievejia

Differential Revision: D61156780

fbshipit-source-id: 2f73bc1a8b436cadde02219859e3560928b55abc