+2 treat classes extending Any as mocks
Summary:
This diff implements some rudimentary duck typing behavior for classes that extend Any. The way we support this right now is by treating classes which extend Any as mocks, which causes certain classes of errors to be suppressed.

Additionally, we add a definition for Any in missingFromStubs.ml, since it's a special form. Without it, classes which only extend Any will not work.

Reviewed By: migeed-z

Differential Revision: D61026859

fbshipit-source-id: 695523cfe1bd51a7784d45028af84d8b65c973a4