Use better and consistent name for `generic_parameters_as_variables`
Summary:
We ran into name collisions between this and the `variables` callback passed
to `Type.create` in recent refactors. In addition, the name wasn't very
descriptive and the inconsistency between the name in `AttributeResolution`
vs in the environment was confusing.

Let's use the same name everywhere, and make it clear in the name that this api
is specifically about unpacking type vars used in a generic class.

Reviewed By: grievejia

Differential Revision: D61161720

fbshipit-source-id: 73622bbbc2dcfc74d2e4f5590c098e193d12dcc0