Simplify: `is_protocol` does not need assumptions
Summary:
My best guess is that at some point in the distant past Pyre used
`protocol_assumptions` as part of the `is_protocol` query, but this
has not been the case in any remotely recent iteration of Pyre - we
determine whether a class is a protocol by looking up the ClassSummary
and asking if it directly inherits `Protocol`.

As a result, the assumptions were pure dead code - we were literally
creating functions *everywhere* that took assumptions as an argument, and not a
single implementation actually used them. Cutting this simplifies the code,
and is probably even a tiny efficiency win since it cost us a closure to do this.

More importantly, the fact that we were passing protocol assumptions made it
impossible to *skim* the code and see what those assumptions actually do; it
*looked* from any skim like they are used to decide what is a protocol, whereas
in reality they are nothing but a process-local cache for solved protocol
parameters.

Reviewed By: grievejia

Differential Revision: D61212961

fbshipit-source-id: 44161eabd507752792872aeddd59f57dc1ac0aee