do not treat dataclass callable attrs as bound methods unless they have a self param
Summary:
Context: https://fb.workplace.com/groups/pyreqa/permalink/8197419127014514/

When we "instantiate" an attribute (whatever that means), if it's not accessed through the class and it's a callable we treat it as a bound method instead of a regular callable.

That means that we assume the first parameter is self, and when we call it we partially apply it to remove the self param before treating it as a callable with the remaining arguments.

This appears to be correct* for callable attrs defined for regular classes, but not for dataclasses where callable attrs are not bound methods at runtime. This difference can be observed by executing the snippet in the diff comments.

I couldn't figure out a way to differentiate between methods and attrs initialized w/ callable values, so for now I just assume that anything with a self param is a method, and will be handled as a bound method. Anything in a dataclass without the self param is now handled as a regular callable.

This diff only affects dataclasses.

Reviewed By: migeed-z

Differential Revision: D61149228

fbshipit-source-id: 0cd10ab307548b3d300cab52f11b21f14207e37f