Do not allow calls to be treated as aliases
Summary:
In the Python typing spec, a call can never be a valid type annotation,
and therefore it can never be a valid type alias.

Pyre does *internally* allow calls to be parsed as types because of our
named function handling logic (it may be that only unit tests actually
need this, I'm unsure) but that is not user-facing because those type
expressions are meaningless in the typing spec.

As a result, it is both possible and desirable to explicitly ban
any `Call` expression from being interpreted as a type alias - such
an expression can potentially define a legacy-style type *variable*,
but never an alias.

In the process we also have fully separated the outer-level match
statements for aliases vs variable declarations.

Note that this change is only really possible because we finished
undoing the eager lowering of the AST into dunder calls :)

Reviewed By: grievejia

Differential Revision: D61296380

fbshipit-source-id: 683a1ba583eb9f02fab225725f53b781cb3e968a