Remove indirection in `Solution.t`
Summary:
Based on a comment, it looks like many years ago there was a thought that maybe
`ConstraintsSet.Solution` would someday generalize `TypeConstraints.Solution`,
and as a result we had a layer of indirection (in just the `.ml` file, not `.mli`).

But (a) this code is already tricky to read both for inherent complexity and
implementation complexity such as redirections and (b) at the moment our team
has limited understanding of the existing code, so making the interfaces look like
some past idea of a possible future isn't useful; if we do decide to generalize we're
going to have to start from scratch building our mental models anyway.

So let's just get rid of this redirection, simplifying a bit.

Reviewed By: kinto0

Differential Revision: D61284514

fbshipit-source-id: 1f64da175c3e8e406d8675d293e6465750309aa6