infer enum attributes with leading but not trailing single underscore as members
Summary:
I misread this part of the spec:

> names that start and end with a single underscore are reserved by enum and cannot be used

The current logic ignores attrs with a single leading underscore, but we should check for the trailing undescore as well.

This affects handling for private attrs (like `__X`), since they get mangled to a single leading underscore. However, their behavior is currently incorrect so this change shouldn't be a big deal.

Reviewed By: stroxler

Differential Revision: D61277701

fbshipit-source-id: a7b84feb8f8e69f7840d7d04c132229e79954ebb