Back out passing variable_map to metaclass
Summary:
We want to pass variable_map whenever the scoped type variables
would come to us from the context in which we are analyzing code.

The `metaclass` function is not an example of this because when this function
runs we are asking "what is the metaclass of class <name>".

The only type vars that would matter for that question are the ones
bound on `<name>` or its superclasses, which means we should not
be passing but rather computing the scoped type vars here.

NOTE: the `resolve_define` and `resolve_literal` funcitons are more complex.
Their low-level implementations do in fact need to take scope as input because
they can both be called from inside of attribute table construction, but
they should *not* take scope as an input in the "global-level" queries (which
can only resolve module-level literals and defines). We'll need to tweak the
APIs in a follow-up diff, but this one is the more urgent because it impacts
caching strategies.

Reviewed By: yangdanny97

Differential Revision: D61307586

fbshipit-source-id: 22fb34d66e07ef5ba252143124cadcddf41baca3