Rename `ConcreteVariable` to `RequriesIterable`
Summary:
I'm not really sure what the original name here was inspired by, but this
error means *exactly* that a type appeared somewhere where an iterable needed,
and it wasn't iterable.

The name ought to reflect that.

(This is motivated because I was utterly baffled by D61230434 in code review,
because with the existing name `ConcreteVariable` it's super unclear what's happening,
and I wasted almost 10 minutes figurin out that this error is 100% about iterables)

Reviewed By: yangdanny97

Differential Revision: D61309085

fbshipit-source-id: 8e337b72a0fb2651febe78b31792d378c83ee43c